zsh: change the prompt char in vicmd to match the one in viins
Dunno if this is a bug in p10k or not but when I switch from
visual to command and then to insert the char was stuck to the one
defined in vicmd.

Dealing with 2 chars fix the problem and anyway I get visual feedback
with the cursor which is a block in command and beam in insert.